requirements.txt
psycopg2-binary
pygame
python-dotenv